packages/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "slowapi>=0.1.9",
    "pypdf>=6.7.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "sqlalchemy[asyncio]>=2.0.47",
    "uvicorn>=0.41.0",
    "pgvector>=0.3.6",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "respx>=0.21.0",
    "faker>=30.0.0",
    "testcontainers[postgres]>=4.0.0",
    "psycopg2-binary>=2.9.11",
]
